% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_data}
\alias{filter_data}
\title{Filter data by countries and cases types}
\usage{
filter_data(data, country_name)
}
\arguments{
\item{data}{A data frame containing a \code{country}  variable.}

\item{country_name}{The country let users to select.}
}
\description{
A function to filter the countries .
}
