% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Countryselection.R
\name{select_countryInput}
\alias{select_countryInput}
\title{Selection of provinces for users in app}
\usage{
select_countryInput(data)
}
\arguments{
\item{data}{A dataset containing \code{Country} in columns.}
}
\description{
Let users to choose which country they want to select in shiny app
}
