% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summary_df}
\alias{summary_df}
\title{#' The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Dataset
The coronavirus provides a tidy format dataset of the 2019 Novel Coronavirus COVID-19 (2019-nCoV) epidemic.}
\format{
A data frame with 7 variables.
\describe{
\item{date}{Date in YYYY-MM-DD format which from 2020-01-22 to 2020-10-07. }
\item{province}{The province or state, when applicable.}
\item{country}{The country or region name}
\item{lat}{Latitude.}
\item{long}{Longitude.}
\item{type}{the type of case (i.e.recover, confirmed, death).}
\item{cases}{The daily cases.}
\item{code}{The code.}
}
}
\source{
The raw data pulled from the Johns Hopkins University Center for Systems Science and Engineering (JHU CCSE) Coronavirus repository.
\href{https://systems.jhu.edu/research/public-health/ncov/}{website}.
}
\usage{
summary_df
}
\description{
#' The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Dataset
The coronavirus provides a tidy format dataset of the 2019 Novel Coronavirus COVID-19 (2019-nCoV) epidemic.
}
\keyword{COVID19,}
\keyword{coronavirus,}
\keyword{datasets,}
\keyword{worldwide}
