% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb}
\alias{tb}
\title{The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Dataset from 2020-01-22 to 2020-10-07.
Changes in the number of confirmed cases in Brazil, the United States, and India}
\format{
A data frame with 7 variables.
\describe{
\item{date}{Date in YYYY-MM-DD format which from 2020-01-22 to 2020-10-07. }
\item{province}{.}
\item{country}{Brazil, the United States, and India.}
\item{lat}{Latitude of these country.}
\item{long}{Longitude of these country.}
\item{type}{the type of case (confirmed).}
\item{cases}{The daily cases.}
}
}
\source{
The raw data pulled from the Johns Hopkins University Center for Systems Science and Engineering (JHU CCSE) Coronavirus repository.
\href{https://systems.jhu.edu/research/public-health/ncov/}{website}.
}
\usage{
tb
}
\description{
The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Dataset from 2020-01-22 to 2020-10-07.
Changes in the number of confirmed cases in Brazil, the United States, and India
}
\keyword{COVID19,}
\keyword{States,India,comfirmed}
\keyword{United}
\keyword{coronavirus,}
\keyword{datasets,}
\keyword{the}
\keyword{worldwide.Brazil,}
